# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MyGoatHerd API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goats:
    get:
      operationId: listGoats
      tags: [goats]
      summary: List all dairy goats
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [dead, first-freshener, leased, on-farm, retired, sold]
        - name: sex
          in: query
          schema:
            type: string
            enum: [doe, buck, wether]
      responses:
        "200":
          description: List of dairy goats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goat"
    post:
      operationId: createGoat
      tags: [goats]
      summary: Add a new dairy goat to the herd
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoatBody"
      responses:
        "201":
          description: Created goat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goat"
  /goats/{id}:
    get:
      operationId: getGoat
      tags: [goats]
      summary: Get a single dairy goat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A dairy goat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goat"
        "404":
          description: Goat not found
    put:
      operationId: updateGoat
      tags: [goats]
      summary: Update a dairy goat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoatBody"
      responses:
        "200":
          description: Updated goat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goat"
        "404":
          description: Goat not found
    delete:
      operationId: deleteGoat
      tags: [goats]
      summary: Remove a dairy goat from the herd
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Goat removed
        "404":
          description: Goat not found
  /goats/import:
    post:
      operationId: importGoats
      tags: [goats]
      summary: Bulk import goats into the herd
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportGoatsBody"
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportGoatsResult"
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with a username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in; returns the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /farms/register:
    post:
      operationId: registerFarm
      tags: [farms]
      summary: Self-service registration of a new farm with its first admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterFarmBody"
      responses:
        "201":
          description: The created farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Farm slug or username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /superadmin/farms:
    get:
      operationId: listFarms
      tags: [superadmin]
      summary: List all farms (superadmin only)
      responses:
        "200":
          description: List of farms with user and goat counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuperadminFarm"
    post:
      operationId: createFarm
      tags: [superadmin]
      summary: Create a new farm with its first admin user (superadmin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmBody"
      responses:
        "201":
          description: The created farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Farm slug or username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /auth/password:
    put:
      operationId: changeOwnPassword
      tags: [auth]
      summary: Change the current user's own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "204":
          description: Password changed
        "400":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Current password is incorrect or not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
...
    Goat:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        registeredName:
          type: string
        adgaId:
          type: string
        damName:
          type: string
        sireName:
          type: string
        maternalGranddamName:
          type: string
        maternalGrandsireName:
          type: string
        paternalGranddamName:
          type: string
        paternalGrandsireName:
          type: string
        damRegNo:
          type: string
        sireRegNo:
          type: string
        maternalGranddamRegNo:
          type: string
        maternalGrandsireRegNo:
          type: string
        paternalGranddamRegNo:
          type: string
        paternalGrandsireRegNo:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        sex:
          type: string
          enum: [doe, buck, wether]
        breed:
          type: string
          enum: [alpine, angora, boer, guernsey, kiko, lamancha, mixed, myotonic, nigerian-dwarf, nubian, oberhasli, pygmy, recorded-grade, saanen, sable, savanna, spanish, texmaster, toggenburg]
        status:
          type: string
          enum: [healthy, watch, treatment, dry]
        milkPerDay:
          type: number
          minimum: 0
          maximum: 10
        lactationStatus:
          type: string
          nullable: true
          enum: [milking, dry, exposed, serviced, pregnant, kid, retired]
        age:
          type: integer
          description: Calculated age in years from dateOfBirth
        description:
          type: string
        imageUrl:
          type: string
        leasedBuck:
          type: boolean
          description: Buck is on a breeding lease and excluded from herd totals
        herdStatus:
          type: string
          nullable: true
          enum: [dead, first-freshener, leased, on-farm, retired, sold]
          description: Current standing of the goat in the herd
        rightEarTattoo:
          type: string
          maxLength: 4
        leftEarTattoo:
          type: string
          maxLength: 4
        rightTailTattoo:
          type: string
          maxLength: 4
        leftTailTattoo:
          type: string
          maxLength: 4
        centerTailTattoo:
          type: string
          maxLength: 8
        eidNumber:
          type: string
          maxLength: 50
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - breed
        - status
        - milkPerDay
        - age
        - createdAt
        - updatedAt
    CreateGoatBody:
      type: object
      properties:
        name:
          type: string
        registeredName:
          type: string
        adgaId:
          type: string
        damName:
          type: string
        sireName:
          type: string
        maternalGranddamName:
          type: string
        maternalGrandsireName:
          type: string
        paternalGranddamName:
          type: string
        paternalGrandsireName:
          type: string
        damRegNo:
          type: string
        sireRegNo:
          type: string
        maternalGranddamRegNo:
          type: string
        maternalGrandsireRegNo:
          type: string
        paternalGranddamRegNo:
          type: string
        paternalGrandsireRegNo:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        sex:
          type: string
          enum: [doe, buck, wether]
        breed:
          type: string
          enum: [alpine, angora, boer, guernsey, kiko, lamancha, mixed, myotonic, nigerian-dwarf, nubian, oberhasli, pygmy, recorded-grade, saanen, sable, savanna, spanish, texmaster, toggenburg]
        status:
          type: string
          enum: [healthy, watch, treatment, dry]
        milkPerDay:
          type: number
          minimum: 0
          maximum: 10
        lactationStatus:
          type: string
          nullable: true
          enum: [milking, dry, exposed, serviced, pregnant, kid, retired]
        age:
          type: integer
          description: Legacy field, calculated from dateOfBirth
        description:
          type: string
        imageUrl:
          type: string
        leasedBuck:
          type: boolean
        herdStatus:
          type: string
          nullable: true
          enum: [dead, first-freshener, leased, on-farm, retired, sold]
        rightEarTattoo:
          type: string
          maxLength: 4
          nullable: true
        leftEarTattoo:
          type: string
          maxLength: 4
          nullable: true
        rightTailTattoo:
          type: string
          maxLength: 4
          nullable: true
        leftTailTattoo:
          type: string
          maxLength: 4
          nullable: true
        centerTailTattoo:
          type: string
          maxLength: 8
          nullable: true
        eidNumber:
          type: string
          maxLength: 50
          nullable: true
      required:
        - name
        - breed
    UpdateGoatBody:
      type: object
      properties:
        name:
          type: string
        registeredName:
          type: string
        adgaId:
          type: string
        damName:
          type: string
        sireName:
          type: string
        maternalGranddamName:
          type: string
        maternalGrandsireName:
          type: string
        paternalGranddamName:
          type: string
        paternalGrandsireName:
          type: string
        damRegNo:
          type: string
        sireRegNo:
          type: string
        maternalGranddamRegNo:
          type: string
        maternalGrandsireRegNo:
          type: string
        paternalGranddamRegNo:
          type: string
        paternalGrandsireRegNo:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        sex:
          type: string
          enum: [doe, buck, wether]
        breed:
          type: string
          enum: [alpine, angora, boer, guernsey, kiko, lamancha, mixed, myotonic, nigerian-dwarf, nubian, oberhasli, pygmy, recorded-grade, saanen, sable, savanna, spanish, texmaster, toggenburg]
        status:
          type: string
          enum: [healthy, watch, treatment, dry]
        milkPerDay:
          type: number
          minimum: 0
          maximum: 10
        lactationStatus:
          type: string
          nullable: true
          enum: [milking, dry, exposed, serviced, pregnant, kid, retired]
        age:
          type: integer
          description: Calculated age in years from dateOfBirth
        description:
          type: string
        imageUrl:
          type: string
        leasedBuck:
          type: boolean
        herdStatus:
          type: string
          nullable: true
          enum: [dead, first-freshener, leased, on-farm, retired, sold]
        rightEarTattoo:
          type: string
          maxLength: 4
          nullable: true
        leftEarTattoo:
          type: string
          maxLength: 4
          nullable: true
        rightTailTattoo:
          type: string
          maxLength: 4
          nullable: true
        leftTailTattoo:
          type: string
          maxLength: 4
          nullable: true
        centerTailTattoo:
          type: string
          maxLength: 8
          nullable: true
        eidNumber:
          type: string
          maxLength: 50
          nullable: true
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ImportGoatRow:
      type: object
      properties:
        name:
          type: string
        registeredName:
          type: string
        adgaId:
          type: string
        sex:
          type: string
          enum: [doe, buck, wether]
        breed:
          type: string
          enum: [alpine, angora, boer, guernsey, kiko, lamancha, mixed, myotonic, nigerian-dwarf, nubian, oberhasli, pygmy, recorded-grade, saanen, sable, savanna, spanish, texmaster, toggenburg]
        dateOfBirth:
          type: string
          format: date-time
        damName:
          type: string
        sireName:
          type: string
        maternalGranddamName:
          type: string
        maternalGrandsireName:
          type: string
        paternalGranddamName:
          type: string
        paternalGrandsireName:
          type: string
        rightEarTattoo:
          type: string
          maxLength: 4
        leftEarTattoo:
          type: string
          maxLength: 4
        rightTailTattoo:
          type: string
          maxLength: 4
        leftTailTattoo:
          type: string
          maxLength: 4
        centerTailTattoo:
          type: string
          maxLength: 8
        eidNumber:
          type: string
          maxLength: 50
        lactationStatus:
          type: string
          enum: [milking, dry, exposed, serviced, pregnant, kid, retired]
      required:
        - name
        - breed
    ImportGoatsBody:
      type: object
      properties:
        goats:
          type: array
          items:
            $ref: "#/components/schemas/ImportGoatRow"
      required:
        - goats
    ImportGoatsResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - skipped
        - errors
...
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, owner, farmhand, superadmin]
        farmSlug:
          type: string
          nullable: true
          description: The slug of the farm this user belongs to. Null for platform superadmins.
      required:
        - id
        - username
        - role
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - username
        - password
...
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword
    Farm:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, suspended]
      required:
        - id
        - slug
        - name
        - status
    SuperadminFarm:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userCount:
          type: integer
        goatCount:
          type: integer
      required:
        - id
        - slug
        - name
        - status
        - userCount
        - goatCount
    RegisterFarmBody:
      type: object
      properties:
        farmName:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 3
          maxLength: 32
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8
      required:
        - farmName
        - slug
        - username
        - password
    CreateFarmBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 3
          maxLength: 32
        adminUsername:
          type: string
          minLength: 1
        adminPassword:
          type: string
          minLength: 8
      required:
        - name
        - slug
        - adminUsername
        - adminPassword
...