# lib/api-zod/src/generated/types/createFarmBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyGoatHerd API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFarmBody {
  /** @minLength 1 */
  name: string;
  /**
   * @minLength 3
   * @maxLength 32
   */
  slug: string;
  /** @minLength 1 */
  adminUsername: string;
  /** @minLength 8 */
  adminPassword: string;
}